+++ packages/raw-data-provider/lib/actions/index.ts
@@ -1,0 +1,1 @@
+export { RawDataProviderStateEntityAction } from './rawDataProviderStateEntityAction';