+++ packages/raw-data-provider/lib/index.ts
@@ -1,0 +1,3 @@
+export { AbstractRawDataProvider } from './abstractRawDataProvider';
+export { migrations } from './migrations';
+export { RawDataProviderStateEntity } from './entities';